+++ app/controllers/s3/request_policies_controller.rb
@@ -3,15 +3,6 @@ module S3
     before_action :authenticate_user!
 
     def show
-      upload_form = S3FormConfigurator.new({
-        url: ENV['S3_URL'],
-        bucket_name: ENV['S3_BUCKET'],
-        aws_key: ENV['AWS_ACCESS_KEY_ID'],
-        aws_secret: ENV['AWS_SECRET_ACCESS_KEY'],
-        upload_path: "pending/#{current_user.id}/#{params[:file_prefix]}",
-        content_type: params[:content_type]
-      })
-
       render json: {
         url: upload_form.url,
         access_key_id: upload_form.access_key_id,
@@ -20,6 +11,18 @@ module S3
         signature: upload_form.signature,
         key: upload_form.key
       }
+    end
+
+    private
+
+    def upload_form
+      S3FormConfigurator.new(
+        url: ENV['S3_URL'],
+        bucket_name: ENV['S3_BUCKET'],
+        aws_key: ENV['AWS_ACCESS_KEY_ID'],
+        aws_secret: ENV['AWS_SECRET_ACCESS_KEY'],
+        upload_path: "pending/#{current_user.id}/#{params[:file_prefix]}",
+        content_type: params[:content_type])
     end
   end
 end