+++ CHANGELOG.md
@@ -6,6 +6,9 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ## [Unreleased][unreleased]
 ### Added
 - Ability to diff form based cards, such as the Additional Information card, in versioning mode
+- New component attachment-manager to handle attachments, to be used on used on multiple cards.
+- New component attachment-item to be used within attachment-manager, to handle previews of attachment, replace and deletion
+- New component s3-file-uploader that wraps the jquery fileupload plugin to make direct uploads to Amazon S3
 
 ### Changed
 - NestedQuestions have unique identifiers